adminOnly flags; move some fields to seperate files; remove acceptedIdeas from begroot
