+++ lib/modules/arguments-form-widgets/index.js
@@ -4,6 +4,7 @@ const eventEmitter  = require('../../../events').emitter;
 module.exports = {
   extend: 'apostrophe-widgets',
   label: 'Arguments form',
+  adminOnly: true,
   addFields: [
     {
       name: 'ideaId',
