+++ lib/modules/arguments-widgets/index.js
@@ -4,6 +4,7 @@ module.exports = {
   extend: 'apostrophe-widgets',
   label: 'Arguments',
   alias: 'arguments',
+  adminOnly: true,
   addFields: [
     {
       name: 'ideaId',
@@ -120,7 +121,7 @@ module.exports = {
        return superOutput(widget, options);
   //     return result;
      };
-     
+
      self.addHelpers({
         showLastName: function (type, widget, user) {
             if (type == 'reactions' && (widget.showLastNameForReactions == 'yes' || (widget.showLastNameForReactions == 'adminonly' && user.role == 'admin'))) {
@@ -128,11 +129,11 @@ module.exports = {
             } else if (type == 'arguments' && (widget.showLastNameForArguments == 'yes' || (widget.showLastNameForArguments == 'adminonly' && user.role == 'admin'))) {
                 return user.lastName;
             }
-            
+
             return '';
         }
      });
    },
-   
-   
+
+
 };
