+++ lib/modules/begroot-widgets/views/phase-voting/index.html
@@ -114,6 +114,6 @@
 
 		<br />
 
-		{{ ideasList.showIdeasList(data.widget.acceptedIdeas, fullHost, user, userHasVoted, isAdmin, data) }}
+		{{ ideasList.showIdeasList(data.ideas, fullHost, user, userHasVoted, isAdmin, data) }}
 	</div>
 </div>
