+++ lib/modules/icon-section-widgets/index.js
@@ -3,6 +3,7 @@ const styleSchema = require('../../../config/styleSchema.js').default;
 module.exports = {
     extend:    'apostrophe-widgets',
     label:     'Icon section',
+    adminOnly: true,
     addFields: [
         {
             name:     'title',
