+++ lib/modules/idea-form-widgets/index.js
@@ -22,6 +22,7 @@ const toSqlDatetime = (inputDate) => {
 module.exports = {
   extend: 'map-widgets',
   label: 'Idea form',
+  adminOnly: true,
   addFields: [
    {
       name: 'redirect',
@@ -775,7 +776,7 @@ module.exports = {
        if (req.data.isAdmin && req.body.modBreak) {
          data.modBreak = req.body.modBreak;
        }
-       
+
        if (req.data.isAdmin) {
            data.budget = req.body.budget;
        }
