+++ lib/modules/idea-map-widgets/index.js
@@ -1,6 +1,7 @@
 module.exports = {
   extend: 'map-widgets',
   label: 'Ideas map',
+  adminOnly: true,
   addFields: [
     {
       name: 'displayCtaButton',
