+++ lib/modules/idea-overview-widgets/index.js
@@ -212,6 +212,7 @@ module.exports = {
   extend: 'apostrophe-widgets',
   label: 'Ideas',
   addFields: fields,
+  adminOnly: true,
   construct: function(self, options) {
     options.arrangeFields = (options.arrangeFields || []).concat([
       {
