+++ lib/modules/idea-single-widgets/index.js
@@ -1,72 +1,14 @@
 const rp = require('request-promise');
 const eventEmitter  = require('../../../events').emitter;
 
-const ideaStates = require('../../../config/idea.js').states;
 const extraFields =  require('../../../config/extraFields.js').fields;
-
-const fields = [
-  {
-      name: 'showShareButtons',
-      type: 'boolean',
-      label: 'Display share buttons?',
-      choices: [
-          {
-              value: true,
-              label: "Yes",
-              showFields: ['shareChannelsSelection']
-          },
-          {
-              value: false,
-              label: "No"
-          },
-      ],
-      def: true
-  },
-  {
-    name: 'shareChannelsSelection',
-    type: 'checkboxes',
-    label: 'Select which share buttons you want to display (if left empty all social buttons will be shown)',
-    choices: [
-        {
-            value: 'facebook',
-            label: "Facebook"
-        },
-        {
-            value: 'twitter',
-            label: "Twitter"
-        },
-        {
-            value: 'mail',
-            label: "E-mail"
-        },
-        {
-            value: 'whatsapp',
-            label: "Whatsapp"
-        },
-    ]
-  }
-].concat(
-    ideaStates.map((state) => {
-      return {
-        type: 'string',
-        name: 'label' +  state.value,
-        label: 'Label for photo: ' + state.value,
-      }
-    })
-  )
-  .concat(
-    ideaStates.map((state) => {
-      return {
-        type: 'string',
-        name: 'labelTime' +  state.value,
-        label: 'Labelfor time status: : ' + state.value,
-      }
-    }));
+const fields = require('lib/fields.js');
 
 module.exports = {
   extend: 'map-widgets',
   label: 'Idea single',
   addFields: fields,
+  adminOnly: true,
   construct: function(self, options) {
      let classIdeaId;
