+++ lib/modules/iframe-widgets/index.js
@@ -3,6 +3,7 @@ const styleSchema = require('../../../config/styleSchema.js').default;
 module.exports = {
   extend: 'apostrophe-widgets',
   label: 'Iframe',
+  adminOnly: true,
   addFields: [
     {
       name: 'url',
