+++ lib/modules/main-image-widgets/index.js
@@ -4,6 +4,7 @@ const styleSchema = require('../../../config/styleSchema.js').default;
 module.exports = {
   extend: 'apostrophe-widgets',
   label: 'Main Image',
+  adminOnly: true,
   addFields: [
     {
       name: 'title',
@@ -40,7 +41,7 @@ module.exports = {
 
     const superLoad = self.load;
     self.load = (req, widgets, callback) => {
-
+      
       widgets.forEach((widget) => {
         if (widget.containerStyles) {
           const containerId = styleSchema.generateId();
@@ -54,7 +55,6 @@ module.exports = {
           widget.formattedTextContainerStyles = styleSchema.format(textContainerId, widget.textContainerStyles);
         }
       });
-
 
       return superLoad(req, widgets, callback);
     }
