+++ lib/modules/section-widgets/index.js
@@ -3,6 +3,7 @@ const styleSchema = require('../../../config/styleSchema.js').default;
 module.exports = {
   extend: 'apostrophe-widgets',
   label: 'Column Section',
+  adminOnly: true,
   controls: {
     position: 'bottom-left'
   },
